iprt: More 64-bit MSC warnings.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@14058 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/include/internal/fs.h
@@ -38,9 +38,9 @@
 
 __BEGIN_DECLS
 
-RTFMODE rtFsModeFromDos(RTFMODE fMode, const char *pszName, unsigned cbName);
-RTFMODE rtFsModeFromUnix(RTFMODE fMode, const char *pszName, unsigned cbName);
-RTFMODE rtFsModeNormalize(RTFMODE fMode, const char *pszName, unsigned cbName);
+RTFMODE rtFsModeFromDos(RTFMODE fMode, const char *pszName, size_t cbName);
+RTFMODE rtFsModeFromUnix(RTFMODE fMode, const char *pszName, size_t cbName);
+RTFMODE rtFsModeNormalize(RTFMODE fMode, const char *pszName, size_t cbName);
 bool    rtFsModeIsValid(RTFMODE fMode);
 bool    rtFsModeIsValidPermissions(RTFMODE fMode);
